backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic>=1.10.0
email-validator
python-jose[cryptography]
passlib[bcrypt,argon2]
boto3
torch
torchvision
numpy
pillow
sqlalchemy>=1.4.0
psycopg2-binary
pydantic-settings
python-multipart>=0.0.5
openai
aiosmtplib
jinja2
alembic>=1.8.0